requirements.txt
sympy
numpy
matplotlib
urx
jupyter
pyserial
opencv-contrib-python
pyyaml
PyQt5
